requirements.txt
streamlit>=1.18
pandas
python-dotenv
openai
openpyxl